refactor: 폴더명 변경 (#10)
