+++ src/components/account/TermsBox.tsx
