+++ src/components/account/TermsComponent.tsx
