+++ src/hooks/useCustomAgree.ts
@@ -1,6 +1,6 @@
 import {useRecoilState} from 'recoil'
 import checkState from '../atoms/agreeState'
-import {auth, TLoginParam} from '../api/accountsApi'
+import {auth, TLoginParam} from '../api/accountApi'
 import {setCookie} from '../util/cookieUtil'
 import {Auth} from '../types/Auth'
 import autheState from '../atoms/authState'
