+++ src/pages/account/LoginPage.tsx
@@ -1,5 +1,5 @@
 import Header from '../../components/Header'
-import LoginForm from '../../components/accounts/LoginForm'
+import LoginForm from '../../components/account/LoginForm'
 
 const LoginPage = () => {
   return (
