+++ src/pages/account/RegisterPage.tsx
@@ -1,4 +1,4 @@
-import RegisterForm from '../../components/accounts/RegisterForm'
+import RegisterForm from '../../components/account/RegisterForm'
 import Header from '../../components/Header'
 
 const RegisterPage = () => {
