+++ src/router/MainRouter.tsx
@@ -1,9 +1,9 @@
 import {BrowserRouter, Navigate, Outlet, Route, Routes} from 'react-router-dom'
 import Layout from './Layout'
 import MainPage from '../pages/MainPage'
-import LoginPage from '../pages/accounts/LoginPage'
-import AgreePage from '../pages/accounts/AgreePage'
-import RegisterPage from '../pages/accounts/RegisterPage'
+import LoginPage from '../pages/account/LoginPage'
+import AgreePage from '../pages/account/AgreePage'
+import RegisterPage from '../pages/account/RegisterPage'
 import MyPage from '../pages/mypage/MyPage'
 import ResultPage from '../pages/result/ResultPage'
 import useCustomAccount, {TCustomAccount} from '../hooks/useCustomAccount'
